refactor: update classes & styles, move to components folder, fix paddings
