+++ packages/ui/src/components/va-card/components/va-card-content/VaCardContent.vue
@@ -10,5 +10,4 @@ import { defineComponent } from 'vue'
 export default defineComponent({
   name: 'VaCardContent',
 })
-
 </script>
