+++ packages/ui/src/components/va-card/index.ts
@@ -1,8 +1,9 @@
 import withConfigTransport from '../../services/config-transport/withConfigTransport'
+
 import _VaCard from './VaCard.vue'
-import VaCardContentBase from './VaCardContent.vue'
-import VaCardTitleBase from './VaCardTitle.vue'
-import VaCardActionsBase from './VaCardActions.vue'
+import VaCardContentBase from './components/va-card-content/VaCardContent.vue'
+import VaCardTitleBase from './components/va-card-title/VaCardTitle.vue'
+import VaCardActionsBase from './components/va-card-actions/VaCardActions.vue'
 
 export const VaCardContent = withConfigTransport(VaCardContentBase)
 export const VaCardTitle = withConfigTransport(VaCardTitleBase)
